# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API",
        "description": "Documentation for API.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/transaction/details": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get transactions by email",
                "operationId": "f942e12dd5b5f65be1e73f47c39e9428",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "PortalID": {
                                                "type": "string",
                                                "example": "10726"
                                            },
                                            "UsernameEmail": {
                                                "type": "string",
                                                "example": "iancreightoncr@gmail.com"
                                            },
                                            "ClientName": {
                                                "type": "string",
                                                "example": "Adele Summers"
                                            },
                                            "TransactionType": {
                                                "type": "string",
                                                "example": "Purchase"
                                            },
                                            "PropertyValue": {
                                                "type": "string",
                                                "example": "5000000"
                                            },
                                            "BaseFee": {
                                                "type": "string",
                                                "example": "27480"
                                            },
                                            "TotalFee": {
                                                "type": "string",
                                                "example": "542170"
                                            },
                                            "QuoteGeneratedDate": {
                                                "type": "string",
                                                "example": "2025-03-27"
                                            },
                                            "InstructedFlag": {
                                                "type": "string",
                                                "example": "0"
                                            },
                                            "InstructedDate": {
                                                "type": "string",
                                                "example": "null"
                                            },
                                            "InstructedMatterNum": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "QuoteID": {
                                                "type": "string",
                                                "example": "2498"
                                            },
                                            "CompletedDate": {
                                                "type": "string",
                                                "example": "null"
                                            },
                                            "CancelledDate": {
                                                "type": "string",
                                                "example": "null"
                                            },
                                            "AddText1": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "AddText2": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "AddText3": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "AddDate1": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "AddDate2": {
                                                "type": "string",
                                                "example": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}